fix: address code review findings across codebase
- Fix reciprocal visibility inheriting source instead of hardcoded "public"
- Fix latestRating null check (=== null → == null) for pending queries
- Fix companion note-only reviews being hidden in watch-event-card
- Fix watchlist-item-card race condition (handlePostReview in onSuccess)
- Fix watchedCount distinct by (tmdbId, mediaType) pair
- Fix duplicate inviteOpen state in watchlist detail page
- Fix duplicate visibility property in watch-event create input
- Fix self-referential FK type with AnyPgColumn (resolves TS cascade)
- Fix backfillPosterData type mismatch by mapping to expected shape
- Fix ratingDistribution to filter by visibility for non-self profiles
- Fix tracker ReviewModal to only trigger removal on event creation
- Fix notification setActionTaken timing (after mutation success)
- Fix watchlist removal dialog to await keepInWatchlist before closing
- Fix watchlist ownership validation (all-or-nothing check)
- Fix keptInWatchlist filter in removal candidates query
- Tighten episode-tracker scope input to z.enum
- Tighten genreId input to require integers
- Wire up infinite scroll load-more in profile activity tab
- Remove unused imports and variables for clean lint/tsc
- Add title wrapping on now-showing marquee
- Fix star ratings link to title route instead of tracker
- Fix spec doc field name (reviewText → note)
- Promote plan doc headings for MD001 compliance

+++ src/integrations/trpc/routers/taste-profile.ts
@@ -310,7 +310,7 @@ export const tasteProfileRouter = {
 		}),
 
 	updateFavouriteGenre: protectedProcedure
-		.input(z.object({ genreId: z.number().nullable() }))
+		.input(z.object({ genreId: z.number().int().nullable() }))
 		.mutation(async ({ input, ctx }) => {
 			await db
 				.update(user)
